(in-progress) Fix broken codebuild
