+++ run_create_codebuild.py
@@ -1,9 +1,7 @@
 #!/usr/bin/env python3
-import time
 
 from env import env
 from run_common import AWSCli
-from run_common import print_message
 from run_common import print_session
 from run_common import reset_template_dir
 from run_create_codebuild_cron import run_create_codebuild_cron
@@ -19,7 +17,6 @@ if __name__ == "__main__":
     args = parse_args()
 
 aws_cli = AWSCli()
-
 
 ################################################################################
 #
